backend/requirements.txt
# ==============================================================================
# MenaCraft Backend — Complete Dependencies
# ==============================================================================

# Core API
fastapi
uvicorn[standard]
pydantic
python-multipart
python-dotenv

# Database & Storage
supabase
httpx

# Network & Scraping
requests
beautifulsoup4
python-whois
validators
google-search-results
tldextract

# AI & NLP Forensics
groq
transformers
torch
torchvision
realitydefender
sentence-transformers
scikit-learn
langdetect
spacy

# Image Forensics
pillow
opencv-python
numpy
ImageHash
piexif

# Video Forensics
yt-dlp
av

# Document Forensics
pypdfium2
python-docx

# Config & Serialization
pyyaml

# Training (optional — only needed for model training)
# tqdm
# wandb